@@ -23,21 +23,21 @@
 		}
 	},
 	"require": {
-		"php": "7.4.*||8.0.*||8.1.*||8.2.*",
-		"typo3/cms-core": "11.5.*",
-		"apache-solr-for-typo3/solr": "^11.5"
+		"php": "8.0.*||8.1.*||8.2.*",
+		"typo3/cms-core": "11.5.*||12.4.*",
+		"apache-solr-for-typo3/solr": "^11.5||^12"
 	},
 	"suggest": {
 		"hmmh/solr-file-indexer": "^2.5"
 	},
 	"require-dev": {
-		"typo3/cms-backend": "^11.5",
-		"typo3/cms-frontend": "^11.5",
-		"typo3/cms-tstemplate": "^11.5",
-		"typo3/cms-install": "^11.5",
+		"typo3/cms-backend": "^11.5||^12.4",
+		"typo3/cms-frontend": "^11.5||^12.4",
+		"typo3/cms-tstemplate": "^11.5||^12.4",
+		"typo3/cms-install": "^11.5||^12.4",
 		"friendsoftypo3/phpstan-typo3": "^0.9",
 		"vimeo/psalm": "^4.10",
-		"typo3/cms-lowlevel": "^11.5",
+		"typo3/cms-lowlevel": "^11.5||^12.4",
 		"helhum/typo3-console": "*",
 		"typo3/coding-standards": "^0.5",
 		"hmmh/solr-file-indexer": "^2.5"
